+++ KeeForge/Views/AcknowledgmentsView.swift
@@ -32,6 +32,16 @@ struct AcknowledgmentsView: View {
             } footer: {
                 Text("MIT License")
             }
+
+            Section {
+                Text(msalLicense)
+                    .font(.caption)
+                    .monospaced()
+            } header: {
+                Text("Microsoft Authentication Library (MSAL)")
+            } footer: {
+                Text("MIT License")
+            }
         }
         .navigationTitle("Acknowledgments")
         .navigationBarTitleDisplayMode(.inline)
@@ -77,6 +87,29 @@ struct AcknowledgmentsView: View {
 
     private let swiftyDropboxLicense = """
         Copyright (c) 2015-2021 Dropbox Inc., http://www.dropbox.com/
+
+        Permission is hereby granted, free of charge, to any person obtaining \
+        a copy of this software and associated documentation files (the \
+        "Software"), to deal in the Software without restriction, including \
+        without limitation the rights to use, copy, modify, merge, publish, \
+        distribute, sublicense, and/or sell copies of the Software, and to \
+        permit persons to whom the Software is furnished to do so, subject to \
+        the following conditions:
+
+        The above copyright notice and this permission notice shall be \
+        included in all copies or substantial portions of the Software.
+
+        THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, \
+        EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF \
+        MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND \
+        NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE \
+        LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION \
+        OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION \
+        WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+        """
+
+    private let msalLicense = """
+        Copyright (c) Microsoft Corporation
 
         Permission is hereby granted, free of charge, to any person obtaining \
         a copy of this software and associated documentation files (the \